requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
python-dotenv==0.19.2
pandas-ta==0.3.14b0
requests>=2.31.0
pandas>=1.3.0
charset-normalizer>=3.2.0
tradingview-ta==3.3.0
websockets==11.0.3
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0